requirements.txt
albumentations==0.4.3
anykeystore==0.2
apex==0.9.10.dev0
attrs==20.2.0
backcall==0.1.0
beautifulsoup4==4.6.3
certifi==2019.11.28
chardet==3.0.4
cliff==2.8.3
cmd2==1.3.9
colorama==0.4.3
configparser==5.0.0
cryptacular==1.5.5
cssselect==1.0.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
efficientnet-pytorch==0.5.1
graphviz==0.14
hupper==1.10.2
idna==2.9
imageio==2.6.1
imgaug==0.2.6
importlib-metadata==1.7.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
joblib==0.14.1
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
lxml==4.0.0
MarkupSafe==1.1.1
matplotlib==3.1.2
MechanicalSoup==0.8.0
MedPy==0.4.0
munch==2.5.0
networkx==2.4
nibabel==3.0.1
numpy==1.18.1
oauthlib==3.1.0
opencv-python-headless==4.1.2.30
pandas==0.25.3
parso==0.5.2
PasteDeploy==2.1.1
pbkdf2==1.3
pbr==5.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plaster==1.0
plaster-pastedeploy==0.7
pretrainedmodels==0.7.4
prettytable==0.7.2
progressbar2==3.34.3
prompt-toolkit==3.0.2
protobuf==3.11.3
ptyprocess==0.6.0
pydicom==1.3.0
Pygments==2.5.2
pyparsing==2.4.6
pyperclip==1.8.0
pyramid==1.10.4
pyramid-mailer==0.15.1
python-dateutil==2.8.1
python-slugify==4.0.1
python-utils==2.4.0
python3-openid==3.2.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
repoze.sendmail==4.4.1
requests==2.23.0
requests-oauthlib==1.3.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
segmentation-models-pytorch==0.1.0
SimpleITK==1.2.4
six==1.13.0
sklearn==0.0
slugify==0.0.1
SQLAlchemy==1.3.20
stevedore==3.2.2
tensorboardX==1.7
text-unidecode==1.3
torch==1.1.0
torchvision==0.3.0
torchviz==0.0.1
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
transaction==3.0.0
translationstring==1.4
urllib3==1.24.3
velruse==1.1.1
venusian==3.0.0
wcwidth==0.1.8
WebOb==1.8.6
WTForms==2.3.3
wtforms-recaptcha==0.3.2
zipp==3.1.0
zope.deprecation==4.4.0
zope.interface==5.1.2
zope.sqlalchemy==1.3
